+++ mayan/apps/common/classes.py
@@ -72,16 +72,6 @@ class ErrorLogNamespace(object):
         return ErrorLogEntry.objects.filter(namespace=self.name)
 
 
-class FakeStorageSubclass(object):
-    """
-    Placeholder class to allow serializing the real storage subclass to
-    support migrations.
-    """
-
-    def __eq__(self, other):
-        return True
-
-
 class MissingItem(object):
     _registry = []