+++ mayan/apps/common/utils.py
@@ -1,9 +1,6 @@
 from __future__ import unicode_literals
 
 import logging
-import os
-import shutil
-import tempfile
 
 from django.conf import settings
 from django.core.exceptions import FieldDoesNotExist
@@ -13,7 +10,6 @@ from django.urls.base import get_script_prefix
 from django.utils.datastructures import MultiValueDict
 from django.utils.http import urlencode as django_urlencode
 from django.utils.http import urlquote as django_urlquote
-from django.utils.module_loading import import_string
 from django.utils.six.moves import reduce as reduce_function
 from django.utils.six.moves import xmlrpc_client
 
@@ -21,7 +17,6 @@ import mayan
 
 from .exceptions import NotLatestVersion, UnknownLatestVersion
 from .literals import DJANGO_SQLITE_BACKEND, MAYAN_PYPI_NAME, PYPI_URL
-from .settings import setting_temporary_directory
 
 logger = logging.getLogger(__name__)
 
@@ -40,52 +35,12 @@ def check_version():
             raise NotLatestVersion(upstream_version=versions[0])
 
 
-# http://stackoverflow.com/questions/123198/how-do-i-copy-a-file-in-python
-def copyfile(source, destination, buffer_size=1024 * 1024):
-    """
-    Copy a file from source to dest. source and dest
-    can either be strings or any object with a read or
-    write method, like StringIO for example.
-    """
-    source_descriptor = get_descriptor(source)
-    destination_descriptor = get_descriptor(destination, read=False)
-
-    while True:
-        copy_buffer = source_descriptor.read(buffer_size)
-        if copy_buffer:
-            destination_descriptor.write(copy_buffer)
-        else:
-            break
-
-    source_descriptor.close()
-    destination_descriptor.close()
-
-
 def encapsulate(function):
     # Workaround Django ticket 15791
     # Changeset 16045
     # http://stackoverflow.com/questions/6861601/
     # cannot-resolve-callable-context-variable/6955045#6955045
     return lambda: function
-
-
-def fs_cleanup(filename, file_descriptor=None, suppress_exceptions=True):
-    """
-    Tries to remove the given filename. Ignores non-existent files
-    """
-    if file_descriptor:
-        os.close(file_descriptor)
-
-    try:
-        os.remove(filename)
-    except OSError:
-        try:
-            shutil.rmtree(filename)
-        except OSError:
-            if suppress_exceptions:
-                pass
-            else:
-                raise
 
 
 def get_related_field(model, related_field_name):
@@ -106,41 +61,6 @@ def get_related_field(model, related_field_name):
         )
 
     return related_field
-
-
-def get_descriptor(file_input, read=True):
-    try:
-        # Is it a file like object?
-        file_input.seek(0)
-    except AttributeError:
-        # If not, try open it.
-        if read:
-            return open(file_input, mode='rb')
-        else:
-            return open(file_input, mode='wb')
-    else:
-        return file_input
-
-
-def get_storage_subclass(dotted_path):
-    """
-    Import a storage class and return a subclass that will always return eq
-    True to avoid creating a new migration when for runtime storage class
-    changes.
-    """
-    imported_storage_class = import_string(dotted_path=dotted_path)
-
-    class StorageSubclass(imported_storage_class):
-        def __init__(self, *args, **kwargs):
-            return super(StorageSubclass, self).__init__(*args, **kwargs)
-
-        def __eq__(self, other):
-            return True
-
-        def deconstruct(self):
-            return ('mayan.apps.common.classes.FakeStorageSubclass', (), {})
-
-    return StorageSubclass
 
 
 def introspect_attribute(attribute_name, obj):
@@ -166,21 +86,6 @@ def introspect_attribute(attribute_name, obj):
             )
     else:
         return attribute_name, obj
-
-
-def TemporaryFile(*args, **kwargs):
-    kwargs.update({'dir': setting_temporary_directory.value})
-    return tempfile.TemporaryFile(*args, **kwargs)
-
-
-def mkdtemp(*args, **kwargs):
-    kwargs.update({'dir': setting_temporary_directory.value})
-    return tempfile.mkdtemp(*args, **kwargs)
-
-
-def mkstemp(*args, **kwargs):
-    kwargs.update({'dir': setting_temporary_directory.value})
-    return tempfile.mkstemp(*args, **kwargs)
 
 
 def resolve(path, urlconf=None):
@@ -275,24 +180,3 @@ def urlquote(link=None, get=None):
         return '%s%s' % (link, django_urlencode(get, doseq=True))
     else:
         return django_urlquote(link)
-
-
-def validate_path(path):
-    if not os.path.exists(path):
-        # If doesn't exist try to create it
-        try:
-            os.mkdir(path)
-        except Exception as exception:
-            logger.debug('unhandled exception: %s', exception)
-            return False
-
-    # Check if it is writable
-    try:
-        fd, test_filepath = tempfile.mkstemp(dir=path)
-        os.close(fd)
-        os.unlink(test_filepath)
-    except Exception as exception:
-        logger.debug('unhandled exception: %s', exception)
-        return False
-
-    return True