+++ mayan/apps/documents/storages.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 from django.utils.module_loading import import_string
 
-from mayan.apps.common.utils import get_storage_subclass
+from mayan.apps.storage.utils import get_storage_subclass
 
 from .settings import (
     setting_documentimagecache_storage,