+++ mayan/apps/sources/tests/test_views.py
@@ -6,13 +6,13 @@ import shutil
 
 from mayan.apps.checkouts.models import NewVersionBlock
 from mayan.apps.common.tests import GenericViewTestCase
-from mayan.apps.common.utils import fs_cleanup, mkdtemp
 from mayan.apps.documents.models import Document
 from mayan.apps.documents.permissions import permission_document_create
 from mayan.apps.documents.tests import (
     TEST_DOCUMENT_DESCRIPTION, TEST_SMALL_DOCUMENT_CHECKSUM,
     TEST_SMALL_DOCUMENT_PATH, GenericDocumentViewTestCase
 )
+from mayan.apps.storage.utils import fs_cleanup, mkdtemp
 
 from ..links import link_upload_version
 from ..literals import SOURCE_CHOICE_WEB_FORM